Refactor AccuWeather logic to async I/O
- Moved logic to logic/AccuWeather.cs
- Replaced synchronous RestSharp call with ExecuteAsync
- Used ILogger instead of Console.WriteLine
- Updated GetWeather function to await the new async method
