+++ GetWeather.cs
@@ -7,7 +7,6 @@ using Microsoft.Azure.WebJobs.Extensions.Http;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
-using RestSharp;
 
 namespace uk.me.timallen.infohub
 {
@@ -25,55 +24,11 @@ namespace uk.me.timallen.infohub
 
             log.LogInformation("location: " + location);
 
-            var response = GetFromAccuweather(location);
+            var response = await AccuWeather.GetForecastAsync(location, log);
 
             log.LogInformation(response);
 
             return new OkObjectResult(response);
-        }
-
-        private static string GetFromAccuweather(string location)
-        {
-            var response = MakeRequest(location);
-            Console.WriteLine(response);
-            dynamic weather = JsonConvert.DeserializeObject(response.Content);
-            return FormatResponse(weather["DailyForecasts"]);
-        }
-
-        private static string FormatResponse(dynamic dailyForecasts)
-        {
-            var forecasts = new System.Collections.Generic.List<object>();
-
-            for(int i = 0; i < dailyForecasts.Count; i++)
-            {
-                dynamic day = dailyForecasts[i];
-                forecasts.Add(new {
-                    day = DateTime.Now.AddDays(i).DayOfWeek.ToString(),
-                    min = day.Temperature.Minimum.Value.ToString(),
-                    max = day.Temperature.Maximum.Value.ToString(),
-                    summary = day.Day.IconPhrase.ToString(),
-                    dayIcon = day.Day.Icon.ToString(),
-                    nightIcon = day.Night.Icon.ToString()
-                });
-            }
-            return JsonConvert.SerializeObject(forecasts);
-        }
-
-        private static IRestResponse MakeRequest(string location)
-        {
-            var client = GetRestClient(location);
-            var request = new RestRequest(Method.GET);
-            return client.Execute(request);
-        }
-
-        private static RestClient GetRestClient(string location)
-        {
-            var serviceApi = Environment.GetEnvironmentVariable("weatherapi");
-            var serviceKey = Environment.GetEnvironmentVariable("weatherapikey");
-            return new RestClient(serviceApi + location +"?apikey="+ serviceKey +"&details=false&metric=true")
-            {
-                Timeout = -1
-            };
         }
     }
 }
